NHL 0.14e

+++ Assets/Scripts/BattleConstants.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 namespace HeroLeft.BattleLogic {
     public static class BattleConstants
     {
-        public const float EnergyCost = 10;
+        public const int EnergyCost = 10;
         public const float ArmorDamageReduction = 10;
 
         public static void CalculateArmor(ref float damage, float armor)
